:truck: refactor: 폴더 구조, 파일 리팩터링
