+++ src/app/(afterLogin)/browse/_component/card-item-skeleton.tsx
