+++ src/app/(afterLogin)/contents/[cid]/_component/preview-contents.tsx
