+++ src/app/(afterLogin)/movies/[id]/_component/card-skeleton.tsx
