+++ src/app/(afterLogin)/movies/[id]/_component/genres.tsx
@@ -1,9 +1,9 @@
 "use client";
 import React, { useEffect, useState } from "react";
-import { Options } from "@/types/ui/types";
 import { getGenres } from "@/services/contents";
-import OptionList from "@/components/ui/option-list";
+import OptionList from "@/app/(afterLogin)/_component/option-list";
 import { useRouter } from "next/navigation";
+import { Options } from "@/model/category";
 
 export default function Genres({ params }: { params: string }) {
   const router = useRouter();
