+++ src/app/(beforeLogin)/signup/_component/icon/index.tsx
