--- src/mocks/browser.ts
@@ -1,6 +1,0 @@
-import { setupWorker } from "msw/browser";
-import { handlers } from "./handlers";
-
-const worker = setupWorker(...handlers);
-
-export default worker;
