--- src/types/browse/types.ts
@@ -1,31 +1,0 @@
-export interface Contents {
-  page: number;
-  media_type: "movie" | "tv";
-  results: Result[];
-  total_pages: number;
-  total_results: number;
-}
-
-export interface Result {
-  adult: boolean;
-  backdrop_path: string;
-  first_air_date: string;
-  genre_ids: number[];
-  id: number;
-  name: string;
-  origin_country: string[];
-  original_language: string;
-  original_title: string;
-  original_name: string;
-  overview: string;
-  popularity: number;
-  poster_path: string;
-  release_date: string;
-  title: string;
-  video: boolean;
-  media_type: string;
-  vote_average: number;
-  vote_count: number;
-}
-
-export interface TvResult {}
