--- src/types/signup/types.ts
@@ -1,6 +1,0 @@
-export interface FormData {
-  email: string;
-  password: string;
-  requiredTerm: boolean;
-  optionsTerm: boolean;
-}
